+++ src/emulator.ts
@@ -17,6 +17,8 @@ import { KillRing } from "./kill-yank/kill-ring";
 import { logger } from "./logger";
 import { MessageManager } from "./message";
 import { PrefixArgumentHandler } from "./prefix-argument";
+import { Configuration } from "./configuration/configuration";
+import { MarkRing } from "./mark-ring";
 
 export interface IEmacsCommandRunner {
   runCommand(commandName: string): undefined | Thenable<{} | undefined | void>;
@@ -32,6 +34,9 @@ export class EmacsEmulator implements Disposable, IEmacsCommandRunner, IMarkMode
 
   private commandRegistry: EmacsCommandRegistry;
 
+  public markRing: MarkRing;
+  private prevExchangedMarks: vscode.Position[] | null;
+
   // tslint:disable-next-line:variable-name
   private _isInMarkMode = false;
   public get isInMarkMode() {
@@ -43,6 +48,9 @@ export class EmacsEmulator implements Disposable, IEmacsCommandRunner, IMarkMode
 
   constructor(textEditor: TextEditor, killRing: KillRing | null = null) {
     this.textEditor = textEditor;
+
+    this.markRing = new MarkRing(Configuration.instance.markRingMax);
+    this.prevExchangedMarks = null;
 
     this.prefixArgumentHandler = new PrefixArgumentHandler();
 
@@ -165,6 +173,9 @@ export class EmacsEmulator implements Disposable, IEmacsCommandRunner, IMarkMode
     });
   }
 
+  /**
+   * C-u
+   */
   public universalArgument() {
     this.prefixArgumentHandler.universalArgument();
   }
@@ -181,7 +192,16 @@ export class EmacsEmulator implements Disposable, IEmacsCommandRunner, IMarkMode
     return command.run(this.textEditor, this.isInMarkMode, prefixArgument);
   }
 
+  /**
+   * C-<SPC>
+   */
   public setMarkCommand() {
+    if (this.prefixArgumentHandler.precedingSingleCtrlU()) {
+      // C-u C-<SPC>
+      this.prefixArgumentHandler.cancel();
+      return this.popMark();
+    }
+
     if (this.isInMarkMode && !this.hasNonEmptySelection()) {
       // Toggle if enterMarkMode is invoked continuously without any cursor move.
       this.exitMarkMode();
@@ -218,13 +238,46 @@ export class EmacsEmulator implements Disposable, IEmacsCommandRunner, IMarkMode
     delete this.killYanker;
   }
 
-  public enterMarkMode() {
+  public enterMarkMode(pushMark = true) {
     this._isInMarkMode = true;
 
     // At this moment, the only way to set the context for `when` conditions is `setContext` command.
     // The discussion is ongoing in https://github.com/Microsoft/vscode/issues/10471
     // TODO: How to write unittest for `setContext`?
     vscode.commands.executeCommand("setContext", "emacs-mcx.inMarkMode", true);
+
+    if (pushMark) {
+      this.pushMark();
+    }
+  }
+
+  public pushMark() {
+    this.prevExchangedMarks = null;
+    this.markRing.push(this.textEditor.selections.map((selection) => selection.active));
+  }
+
+  public popMark() {
+    const prevMark = this.markRing.pop();
+    if (prevMark) {
+      this.textEditor.selections = prevMark.map((position) => new Selection(position, position));
+    }
+  }
+
+  public exchangePointAndMark() {
+    const prevMarks = this.prevExchangedMarks || this.markRing.getTop();
+    this.enterMarkMode(false);
+    this.prevExchangedMarks = this.textEditor.selections.map((selection) => selection.active);
+
+    if (prevMarks) {
+      const affectedLen = Math.min(this.textEditor.selections.length, prevMarks.length);
+      const affectedSelections = this.textEditor.selections.slice(0, affectedLen).map((selection, i) => {
+        const prevMark = prevMarks[i];
+        return new vscode.Selection(selection.active, prevMark);
+      });
+      const newSelections = affectedSelections.concat(this.textEditor.selections.slice(affectedLen));
+      this.textEditor.selections = newSelections;
+      this.textEditor.revealRange(this.textEditor.selection);
+    }
   }
 
   public exitMarkMode() {